backend/requirements.txt
fastapi~=0.115.12
uvicorn~=0.34.0
pydantic~=2.11.3
httpx~=0.28.1
python-dotenv~=1.1.0
protobuf~=5.29.4